Add new option 'offsetdays', +some fixes
Adding new optional setting 'offsetdays', in order to compensate for the number of days left until trash has to be pulled out to the street by the owner. The reason for this is that in cases where trash is taken away during the night or early in the morning, the user has to prepare them outside the day before. Also this prevents displaying some days left information for the day when the trash has already been taken.

If not specified, 'offsetdays' is 0, and the component behaves as before. If the user sets 'offsetdays' to 1 for example, the component will simply substract 1 from the number of days left. It doesn't change the date, it only changes the number of days left.

This makes automations more easy, for example notificatins can be triggered simply without any template works etc. to warn the user that it's the last chance to prepare the trash in time.

Also fixed the missing default icon, the default name and added the data source to the attributes.